refactor(messaging): rename NATS subject and associated classes
