+++ backend/api_gateway/src/main/resources/application.yml
@@ -24,5 +24,5 @@ app:
         findAll: jobs.findAllCategories
     messageChannels:
       subjects:
-        findAllOfUser: messaging.findAllUserMessageChannels
-        findOneOfUser: messaging.findOneUserMessageChannel
+        findAllOfUser: messaging.findAllChannelsOfUser
+        findOneOfUser: messaging.findOneChannelOfUser
