+++ backend/messaging/src/main/resources/application.properties
@@ -1,21 +1,21 @@
-spring.cloud.function.definition=findAllUserMessageChannels;findOneUserMessageChannel
+spring.cloud.function.definition=getChannelsOfUser;getChannelOfUser
 
-spring.cloud.stream.bindings.findAllUserMessageChannels-in-0.destination=messaging.findAllUserMessageChannels
-spring.cloud.stream.bindings.findAllUserMessageChannels-in-0.group=messagingSvc
-spring.cloud.stream.bindings.findAllUserMessageChannels-in-0.binder=nats
-spring.cloud.stream.bindings.findAllUserMessageChannels-in-0.content-type=application/vnd.linkedout.proto-request
-spring.cloud.stream.bindings.findAllUserMessageChannels-out-0.destination=
-spring.cloud.stream.bindings.findAllUserMessageChannels-out-0.group=messagingSvc
-spring.cloud.stream.bindings.findAllUserMessageChannels-out-0.binder=nats
-spring.cloud.stream.bindings.findAllUserMessageChannels-out-0.content-type=application/vnd.linkedout.proto-response
+spring.cloud.stream.bindings.getChannelsOfUser-in-0.destination=messaging.findAllChannelsOfUser
+spring.cloud.stream.bindings.getChannelsOfUser-in-0.group=messagingSvc
+spring.cloud.stream.bindings.getChannelsOfUser-in-0.binder=nats
+spring.cloud.stream.bindings.getChannelsOfUser-in-0.content-type=application/vnd.linkedout.proto-request
+spring.cloud.stream.bindings.getChannelsOfUser-out-0.destination=
+spring.cloud.stream.bindings.getChannelsOfUser-out-0.group=messagingSvc
+spring.cloud.stream.bindings.getChannelsOfUser-out-0.binder=nats
+spring.cloud.stream.bindings.getChannelsOfUser-out-0.content-type=application/vnd.linkedout.proto-response
 
-spring.cloud.stream.bindings.findOneUserMessageChannel-in-0.destination=messaging.findOneUserMessageChannel
-spring.cloud.stream.bindings.findOneUserMessageChannel-in-0.group=messagingSvc
-spring.cloud.stream.bindings.findOneUserMessageChannel-in-0.binder=nats
-spring.cloud.stream.bindings.findOneUserMessageChannel-in-0.content-type=application/vnd.linkedout.proto-request
-spring.cloud.stream.bindings.findOneUserMessageChannel-out-0.destination=
-spring.cloud.stream.bindings.findOneUserMessageChannel-out-0.group=messagingSvc
-spring.cloud.stream.bindings.findOneUserMessageChannel-out-0.binder=nats
-spring.cloud.stream.bindings.findOneUserMessageChannel-out-0.content-type=application/vnd.linkedout.proto-response
+spring.cloud.stream.bindings.getChannelOfUser-in-0.destination=messaging.findOneChannelOfUser
+spring.cloud.stream.bindings.getChannelOfUser-in-0.group=messagingSvc
+spring.cloud.stream.bindings.getChannelOfUser-in-0.binder=nats
+spring.cloud.stream.bindings.getChannelOfUser-in-0.content-type=application/vnd.linkedout.proto-request
+spring.cloud.stream.bindings.getChannelOfUser-out-0.destination=
+spring.cloud.stream.bindings.getChannelOfUser-out-0.group=messagingSvc
+spring.cloud.stream.bindings.getChannelOfUser-out-0.binder=nats
+spring.cloud.stream.bindings.getChannelOfUser-out-0.content-type=application/vnd.linkedout.proto-response
 
 nats.spring.server=nats://localhost:4222
